+++ lib/date/index.ts
@@ -4,6 +4,7 @@ import {endOfUTC}       from './endOfUTC';
 import {format}         from './format';
 import {isDateFormat}   from './isFormat';
 import {isDate}         from './is';
+import {isLeap}         from './isLeap';
 import {nowUnix}        from './nowUnix';
 import {nowUnixMs}      from './nowUnixMs';
 import {setTimeUTC}     from './setTimeUTC';
@@ -20,6 +21,7 @@ export {
     isDateFormat,
     isDate,
     isDate as is,
+    isLeap,
     nowUnix,
     nowUnixMs,
     setTimeUTC,